Add deterministic signal timeline core
